.requirements.txt
torch==2.7.0
ultralytics>=8.0.0
opencv-python>=4.7.0
pyserial>=3.5
numpy>=1.22.0
python-dotenv